Fix MingGW compiler warnings
reg. comparison of integer expressions of different signedness: 'int' and 'size_t' {aka 'long long unsigned int'} (-Wsign-compare)

+++ src/engraving/libmscore/fret.cpp
@@ -131,9 +131,9 @@ EngravingItem* FretDiagram::linkedClone()
 std::shared_ptr<FretDiagram> FretDiagram::createFromString(Score* score, const String& s)
 {
     auto fd = Factory::makeFretDiagram(score->dummy()->segment());
-    size_t strings = s.size();
+    int strings = static_cast<int>(s.size());
 
-    fd->setStrings(static_cast<int>(strings));
+    fd->setStrings(strings);
     fd->setFrets(4);
     fd->setPropertyFlags(Pid::FRET_STRINGS, PropertyFlags::UNSTYLED);
     fd->setPropertyFlags(Pid::FRET_FRETS,   PropertyFlags::UNSTYLED);
